server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
httpx==0.27.2
pydantic==2.8.2
redis==5.0.7
python-dotenv==1.0.1
tenacity==9.0.0
pytest==8.3.2
pytest-vcr==1.0.2
vcrpy==6.0.1
gunicorn==22.0.0
